These initial changes allow CMake to generate an Xcode project that understands Swift files.
I'm told the change in cmGlobalXCodeGenerator.cxx is good but adding
the swift extension to CMakeCCompiler.cmake.in is kind of a hack.
Instead we should write a full language generator.

For now this works. In future commits, this will be backed out in favor
of a new language generator.