Swift/Makefile: Added initial support for add_library.
Allows intermixed Swift and C files.
This will generate a library (presumably dynamic) that can be linked to.
This does not correctly generate the partial .swiftmodules and merge
them. The partials are generated, with a slightly incorrect (wrt Xcode)
    file name, but no merging is done because this requires CMake to
    understand that it needs to run another step with new commands.
However, this process is good enough to produce a library that can be
called with C linkage. This is extremely valuable for the base case on
Android which must always build NDK stuff into a dynamic library and be
invoked via JNI.